+++ server/utils/crmFieldFormatter.ts
@@ -63,14 +63,3 @@ export const formatPercentage = (value: string): string => {
   if (Number.isNaN(number)) return value
   return `${number}%`
 }
-
-export const splitCamelCase = (value: string): string => {
-  return value.replace(/([a-z])([A-Z])/g, '$1 $2').toLowerCase()
-}
-
-export const removeUnderscore = (value: string): string => {
-  return value
-    .replace(/_/g, ' ')
-    .toLowerCase()
-    .replace(/\b\w/g, char => char.toUpperCase())
-}
