+++ server/utils/utils.ts
@@ -42,3 +42,14 @@ export const currentIsoDate = (): string => {
 export const isNotEmpty = (value: string): boolean => {
   return value !== undefined && value !== null && value !== ''
 }
+
+export const splitCamelCase = (value: string): string => {
+  return value.replace(/([a-z])([A-Z])/g, '$1 $2').toLowerCase()
+}
+
+export const removeUnderscore = (value: string): string => {
+  return value
+    .replace(/_/g, ' ')
+    .toLowerCase()
+    .replace(/\b\w/g, char => char.toUpperCase())
+}
